feat: separate hamburger menu into new NavMenu component
