+++ assets/css/app.scss
@@ -32,6 +32,7 @@ $shuttle-picker-width: 12.6rem;
 @import "modal";
 @import "nav";
 @import "nav/left_nav";
+@import "nav/nav_menu";
 @import "nav/top_nav";
 @import "nav/top_nav_mobile";
 @import "nav/bottom_nav_mobile";
