+++ assets/css/nav/_left_nav.scss
@@ -5,7 +5,6 @@
   justify-content: space-between;
   box-sizing: border-box;
   height: 100%;
-  position: relative;
 
   background-color: $color-gray-50;
