+++ assets/css/nav/_nav_menu.scss
@@ -1,0 +1,129 @@
+.m-nav-menu--open {
+  width: 18.125rem;
+  max-width: 100%;
+  height: 100%;
+  position: absolute;
+  top: 0;
+  left: 0;
+  z-index: 110;
+  padding: 0px;
+  background-color: $color-gray-50;
+  border-right: 1px solid $color-gray-300;
+}
+
+.m-nav-menu--closed {
+  display: none;
+}
+
+.m-nav-menu__icon svg {
+  fill: $color-gray-600;
+  &:hover,
+  &:active,
+  &:focus {
+    fill: $color-gray-500;
+  }
+}
+
+.m-nav-menu__header {
+  margin-bottom: 0.5rem;
+  padding: 0.625rem 1.5rem 0.625rem 1.5rem;
+  background-color: $color-gray-100;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.m-nav-menu__link,
+.m-nav-menu__button {
+  text-decoration: none;
+  display: flex;
+  align-items: center;
+  padding: 0.5rem 0.625rem;
+  box-sizing: border-box;
+  width: 100%;
+  height: 2.25rem;
+  border-radius: 0.375rem;
+
+  &:hover {
+    background-color: $color-gray-100;
+    color: $color-gray-700;
+    font-weight: 500;
+  }
+
+  &.m-nav-menu__link--active {
+    background-color: $color-eggplant-100;
+    font-weight: 500;
+    color: $color-eggplant-800;
+  }
+}
+
+.m-nav-menu__close {
+  width: 1rem;
+  height: 1rem;
+  svg {
+    fill: $color-gray-600;
+  }
+}
+
+.m-nav-menu__close-icon {
+  display: block;
+  height: 1rem;
+}
+
+.m-nav-menu__logo {
+  width: 2.5rem;
+  height: 1.5rem;
+
+  svg {
+    width: 100%;
+    height: 100%;
+  }
+}
+
+.m-nav-menu__logo-icon {
+  svg {
+    fill: $color-eggplant-900;
+    max-width: 100%;
+    height: 100%;
+  }
+}
+
+.m-nav-menu__button,
+.m-nav-menu__link {
+  color: $color-gray-600;
+  &:hover,
+  &:active,
+  &:focus {
+    color: $color-gray-500;
+  }
+  height: 3rem;
+  display: flex;
+  padding-top: 0.25rem;
+  padding-bottom: 0.25rem;
+}
+
+.m-nav-menu__icon {
+  padding-left: 1.5rem;
+  padding-right: 1.5rem;
+  svg {
+    width: 1.5rem;
+    height: 1.5rem;
+  }
+}
+
+.m-nav-menu__overlay {
+  position: fixed;
+  padding: 0px;
+  margin: 0px;
+  display: none;
+}
+
+.m-nav-menu__overlay--open {
+  display: block;
+  z-index: 109;
+  background: rgba(88, 88, 88, 0.23);
+  height: 100%;
+  width: 100%;
+  top: 0;
+  left: 0;
+}
