+++ assets/css/nav/_top_nav_mobile.scss
@@ -46,22 +46,7 @@
   align-items: center;
 }
 
-.m-top-nav-mobile__overlay {
-  position: fixed;
-  padding: 0px;
-  margin: 0px;
-  display: none;
-}
-.m-top-nav-mobile__overlay--open {
-  display: block;
-  z-index: 109;
-  background: rgba(88, 88, 88, 0.23);
-  height: 100%;
-  width: 100%;
-}
-
-.m-top-nav-mobile__icon svg,
-.m-top-nav-mobile__menu-icon svg {
+.m-top-nav-mobile__icon svg {
   fill: $color-gray-600;
   &:hover,
   &:active,
@@ -116,99 +101,6 @@
   }
 }
 
-.m-top-nav-mobile__menu--open {
-  width: 18.125rem;
-  max-width: 100%;
-  height: 100%;
-  position: absolute;
-  left: 0;
-  z-index: 110;
-  padding: 0px;
-  background-color: $color-gray-50;
-  border-right: 1px solid $color-gray-300;
-}
-
-.m-top-nav-mobile__menu--closed {
-  display: none;
-}
-
-.m-top-nav-mobile__menu-header {
-  margin-bottom: 0.5rem;
-  padding: 0.625rem 1.5rem 0.625rem 1.5rem;
-  background-color: $color-gray-100;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.m-top-nav-mobile__close {
-  width: 1rem;
-  height: 1rem;
-  svg {
-    fill: $color-gray-600;
-  }
-}
-
-.m-top-nav-mobile__close-icon {
-  display: block;
-  height: 1rem;
-}
-
-.m-top-nav-mobile__logo-icon {
-  svg {
-    fill: $color-eggplant-900;
-    max-width: 100%;
-    height: 100%;
-  }
-}
-
 .m-top-nav-mobile__header-text {
   color: $color-gray-900;
-}
-
-.m-top-nav-mobile__menu-link,
-.m-top-nav-mobile__menu-button {
-  text-decoration: none;
-  display: flex;
-  align-items: center;
-  padding: 0.5rem 0.625rem;
-  box-sizing: border-box;
-  width: 100%;
-  height: 2.25rem;
-  border-radius: 0.375rem;
-
-  &:hover {
-    background-color: $color-gray-100;
-    color: $color-gray-700;
-    font-weight: 500;
-  }
-
-  &.m-top-nav-mobile__menu-link--active {
-    background-color: $color-eggplant-100;
-    font-weight: 500;
-    color: $color-eggplant-800;
-  }
-}
-
-.m-top-nav-mobile__menu-button,
-.m-top-nav-mobile__menu-link {
-  color: $color-gray-600;
-  &:hover,
-  &:active,
-  &:focus {
-    color: $color-gray-500;
-  }
-  height: 3rem;
-  display: flex;
-  padding-top: 0.25rem;
-  padding-bottom: 0.25rem;
-}
-
-.m-top-nav-mobile__menu-icon {
-  padding-left: 1.5rem;
-  padding-right: 1.5rem;
-  svg {
-    width: 1.5rem;
-    height: 1.5rem;
-  }
 }
